// AgendaVirtual/src/app/models/servicios.ts
export interface Servicios {

    tipo_servicio?: string;
    fecha_solicitada?: string;
    hora_solicitada?: string;
    tipo_vehiculo?: string;
    telefono?:string;
    placa?: string;
    conductor?: string;
    numero_ident?: string;
    numero_contacto?: string;
    observaciones?: string;
    estado?: string;
}